+++ frontend/src/app/brand/[query]/page.tsx
@@ -1,23 +1,22 @@
-import GridCardProductCategory from "../components/GridCardProductCategory";
+import GridCardProductBrand from "../components/GridCardProductBrand";
 import Link from "next/link";
 import React from "react";
-import SliderCategory from "@/components/SliderCategory";
-
+import SliderBrand from "@/components/SliderBrand";
 
 const page = ({ params }: any) => {
   return (
     <div>
-      <SliderCategory/>
+      <SliderBrand/>
       <div className="text-[1rem] breadcrumbs">
         <ul>
           <li className="text-pink-500">
             <Link href={"/"}>Home</Link>
           </li>
-          <li>Categoría: {params.query.replace(/%20/g,' ')}</li>
+          <li>Marca: {params.query.replace(/%20/g,' ')}</li>
         </ul>
       </div>
       <div>
-        <GridCardProductCategory category={params.query} />
+        <GridCardProductBrand brand={params.query} />
       </div>
     </div>
   );
